requirements.txt
asgiref==3.7.2
coreapi==2.3.3
coreschema==0.0.4
Django
django-cors-headers
djangorestframework
geopy==2.4.0
gunicorn==21.2.0
Pillow==10.0.1
psycopg2-binary==2.9.7
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
whitenoise==6.5.0
twilio
